+++ src/connectedmachine/azext_connectedmachine/aaz/latest/connectedmachine/license_profile/_create.py
@@ -22,9 +22,9 @@ class Create(AAZCommand):
     """
 
     _aaz_info = {
-        "version": "2024-11-10-preview",
+        "version": "2026-06-16-preview",
         "resources": [
-            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.hybridcompute/machines/{}/licenseprofiles/{}", "2024-11-10-preview"],
+            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.hybridcompute/machines/{}/licenseprofiles/{}", "2026-06-16-preview"],
         ]
     }
 
@@ -50,16 +50,13 @@ class Create(AAZCommand):
             help="The name of the license profile.",
             required=True,
             enum={"default": "default"},
-            fmt=AAZStrArgFormat(
-                pattern="[a-zA-Z0-9-_\.]+",
-            ),
         )
         _args_schema.machine_name = AAZStrArg(
             options=["--machine-name"],
             help="The name of the hybrid machine.",
             required=True,
             fmt=AAZStrArgFormat(
-                pattern="[a-zA-Z0-9-_\.]+",
+                pattern="[a-zA-Z0-9-_\\.]+",
             ),
         )
         _args_schema.resource_group = AAZResourceGroupNameArg(
@@ -74,26 +71,6 @@ class Create(AAZCommand):
             arg_group="EsuProfile",
             help="The resource id of the license.",
         )
-
-        # define Arg Group "Parameters"
-
-        _args_schema = cls._args_schema
-        _args_schema.location = AAZResourceLocationArg(
-            arg_group="Parameters",
-            help="The geo-location where the resource lives",
-            required=True,
-            fmt=AAZResourceLocationArgFormat(
-                resource_group_arg="resource_group",
-            ),
-        )
-        _args_schema.tags = AAZDictArg(
-            options=["--tags"],
-            arg_group="Parameters",
-            help="Resource tags.",
-        )
-
-        tags = cls._args_schema.tags
-        tags.Element = AAZStrArg()
 
         # define Arg Group "ProductProfile"
 
@@ -129,6 +106,26 @@ class Create(AAZCommand):
             help="Indicates the current status of the product features.",
             enum={"Disabled": "Disabled", "Disabling": "Disabling", "Enabled": "Enabled", "Enabling": "Enabling", "Failed": "Failed", "Unknown": "Unknown"},
         )
+
+        # define Arg Group "Resource"
+
+        _args_schema = cls._args_schema
+        _args_schema.location = AAZResourceLocationArg(
+            arg_group="Resource",
+            help="The geo-location where the resource lives",
+            required=True,
+            fmt=AAZResourceLocationArgFormat(
+                resource_group_arg="resource_group",
+            ),
+        )
+        _args_schema.tags = AAZDictArg(
+            options=["--tags"],
+            arg_group="Resource",
+            help="Resource tags.",
+        )
+
+        tags = cls._args_schema.tags
+        tags.Element = AAZStrArg()
 
         # define Arg Group "SoftwareAssurance"
 
@@ -169,7 +166,7 @@ class Create(AAZCommand):
                     session,
                     self.on_200_201,
                     self.on_error,
-                    lro_options={"final-state-via": "azure-async-operation"},
+                    lro_options={"final-state-via": "location"},
                     path_format_arguments=self.url_parameters,
                 )
             if session.http_response.status_code in [200, 201]:
@@ -178,7 +175,7 @@ class Create(AAZCommand):
                     session,
                     self.on_200_201,
                     self.on_error,
-                    lro_options={"final-state-via": "azure-async-operation"},
+                    lro_options={"final-state-via": "location"},
                     path_format_arguments=self.url_parameters,
                 )
 
@@ -225,7 +222,7 @@ class Create(AAZCommand):
         def query_parameters(self):
             parameters = {
                 **self.serialize_query_param(
-                    "api-version", "2024-11-10-preview",
+                    "api-version", "2026-06-16-preview",
                     required=True,
                 ),
             }
@@ -400,7 +397,7 @@ class Create(AAZCommand):
             product_profile.error = AAZObjectType(
                 flags={"read_only": True},
             )
-            _CreateHelper._build_schema_error_detail_read(product_profile.error)
+            _CreateHelper._build_schema_azure_resourcemanager_commontypes_errordetail_read(product_profile.error)
             product_profile.product_features = AAZListType(
                 serialized_name="productFeatures",
             )
@@ -434,7 +431,7 @@ class Create(AAZCommand):
             _element.error = AAZObjectType(
                 flags={"read_only": True},
             )
-            _CreateHelper._build_schema_error_detail_read(_element.error)
+            _CreateHelper._build_schema_azure_resourcemanager_commontypes_errordetail_read(_element.error)
             _element.name = AAZStrType()
             _element.subscription_status = AAZStrType(
                 serialized_name="subscriptionStatus",
@@ -474,62 +471,60 @@ class Create(AAZCommand):
 class _CreateHelper:
     """Helper class for Create"""
 
-    _schema_error_detail_read = None
+    _schema_azure_resourcemanager_commontypes_errordetail_read = None
 
     @classmethod
-    def _build_schema_error_detail_read(cls, _schema):
-        if cls._schema_error_detail_read is not None:
-            _schema.additional_info = cls._schema_error_detail_read.additional_info
-            _schema.code = cls._schema_error_detail_read.code
-            _schema.details = cls._schema_error_detail_read.details
-            _schema.message = cls._schema_error_detail_read.message
-            _schema.target = cls._schema_error_detail_read.target
+    def _build_schema_azure_resourcemanager_commontypes_errordetail_read(cls, _schema):
+        if cls._schema_azure_resourcemanager_commontypes_errordetail_read is not None:
+            _schema.additional_info = cls._schema_azure_resourcemanager_commontypes_errordetail_read.additional_info
+            _schema.code = cls._schema_azure_resourcemanager_commontypes_errordetail_read.code
+            _schema.details = cls._schema_azure_resourcemanager_commontypes_errordetail_read.details
+            _schema.message = cls._schema_azure_resourcemanager_commontypes_errordetail_read.message
+            _schema.target = cls._schema_azure_resourcemanager_commontypes_errordetail_read.target
             return
 
-        cls._schema_error_detail_read = _schema_error_detail_read = AAZObjectType(
+        cls._schema_azure_resourcemanager_commontypes_errordetail_read = _schema_azure_resourcemanager_commontypes_errordetail_read = AAZObjectType(
             flags={"read_only": True}
         )
 
-        error_detail_read = _schema_error_detail_read
-        error_detail_read.additional_info = AAZListType(
+        azure_resourcemanager_commontypes_errordetail_read = _schema_azure_resourcemanager_commontypes_errordetail_read
+        azure_resourcemanager_commontypes_errordetail_read.additional_info = AAZListType(
             serialized_name="additionalInfo",
             flags={"read_only": True},
         )
-        error_detail_read.code = AAZStrType(
+        azure_resourcemanager_commontypes_errordetail_read.code = AAZStrType(
             flags={"read_only": True},
         )
-        error_detail_read.details = AAZListType(
+        azure_resourcemanager_commontypes_errordetail_read.details = AAZListType(
             flags={"read_only": True},
         )
-        error_detail_read.message = AAZStrType(
+        azure_resourcemanager_commontypes_errordetail_read.message = AAZStrType(
             flags={"read_only": True},
         )
-        error_detail_read.target = AAZStrType(
+        azure_resourcemanager_commontypes_errordetail_read.target = AAZStrType(
             flags={"read_only": True},
         )
 
-        additional_info = _schema_error_detail_read.additional_info
+        additional_info = _schema_azure_resourcemanager_commontypes_errordetail_read.additional_info
         additional_info.Element = AAZObjectType()
 
-        _element = _schema_error_detail_read.additional_info.Element
-        _element.info = AAZObjectType(
+        _element = _schema_azure_resourcemanager_commontypes_errordetail_read.additional_info.Element
+        _element.info = AAZAnyType(
             flags={"read_only": True},
         )
         _element.type = AAZStrType(
             flags={"read_only": True},
         )
 
-        details = _schema_error_detail_read.details
-        details.Element = AAZObjectType(
-            flags={"read_only": True},
-        )
-        cls._build_schema_error_detail_read(details.Element)
+        details = _schema_azure_resourcemanager_commontypes_errordetail_read.details
+        details.Element = AAZObjectType()
+        cls._build_schema_azure_resourcemanager_commontypes_errordetail_read(details.Element)
 
-        _schema.additional_info = cls._schema_error_detail_read.additional_info
-        _schema.code = cls._schema_error_detail_read.code
-        _schema.details = cls._schema_error_detail_read.details
-        _schema.message = cls._schema_error_detail_read.message
-        _schema.target = cls._schema_error_detail_read.target
+        _schema.additional_info = cls._schema_azure_resourcemanager_commontypes_errordetail_read.additional_info
+        _schema.code = cls._schema_azure_resourcemanager_commontypes_errordetail_read.code
+        _schema.details = cls._schema_azure_resourcemanager_commontypes_errordetail_read.details
+        _schema.message = cls._schema_azure_resourcemanager_commontypes_errordetail_read.message
+        _schema.target = cls._schema_azure_resourcemanager_commontypes_errordetail_read.target
 
 
 __all__ = ["Create"]