+++ src/datadog/HISTORY.rst
@@ -2,6 +2,15 @@
 
 Release History
 ===============
+3.1.0b1
++++++++
+* Updated to API version 2025-12-26-preview
+* Added command: az datadog activate-saas
+* Added command: az datadog monitor link-saas
+* Added command: az datadog monitor latest-linked-saas
+* Added command: az datadog monitor get-default-application-key
+* Added command: az datadog monitor manage-sre-agent-connector
+
 3.0.0
 ++++++
 * Added multi subscription support