+++ src/discovery/HISTORY.rst
@@ -3,6 +3,10 @@
 Release History
 ===============
 
+1.0.1
+++++++
+* Promote to stable version.
+
 1.0.0b1
 ++++++
 * Initial release.